fix: incorrect polling frequency after ping
When device was offline and turned online, the polling interval is
started again using the polling frequency stated in the settings.
However this was improperly gathered resulting in a polling frequency
which gathered results numerous times each second.

This lead to instability to the app and device remaining unavailable.

Fixes #2